Fix eslint warnings and rename icons
