+++ src/components/atoms/Icon/File.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import IconProps from "./icon";
 
-function Icon({ style, fill }: IconProps) {
+function Icon({ style }: IconProps) {
   return (
     <svg
       style={style}
