+++ src/components/atoms/Icon/MoreOption.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import IconProps from "./icon";
 
-function Icon({ style, fill }: IconProps) {
+function Icon({ style }: IconProps) {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
