+++ src/components/atoms/Icon/PlusCircle.tsx
