+++ src/components/atoms/Icon/University.tsx
